modules/battery-upower: Fix battery detection.
Module was looking at the `NativePath` and matched it for the string
'battery', but this doesn't match:

    'NativePath': <'bq24150a-0'>

Now we check the type and technology, which means we might still match
several batteries; but this is better than matching none.

We might want a whitelist, just as proposed for the battery applet, but
for now, this is good enough.

mce doesn't seem to receive any dbus signals regarding property changes,
however.