battery-upower: extend n900 blacklists
This allows the N900 to once again detect charging in MCE.

We need to use isp1704, not any of the twl* devices. There is no way to
distinguish them by code in upower, so let's just blacklist them.

Also blacklist (yet another) dummy battery, which might just be
rx51-battery under a new name. This is not strictly necessary, since it
-currently- is ordered to appear after the real battery, but let's just
blacklist it anyway.